refactor: (#75) dto package
