+++ simtong-application/src/main/kotlin/team/comit/simtong/domain/auth/usecase/ReissueTokenUseCase.kt
@@ -3,7 +3,7 @@ package team.comit.simtong.domain.auth.usecase
 import team.comit.simtong.domain.auth.exception.RefreshTokenNotFoundException
 import team.comit.simtong.domain.auth.spi.JwtPort
 import team.comit.simtong.domain.auth.spi.QueryRefreshTokenPort
-import team.comit.simtong.domain.auth.usecase.dto.TokenResponse
+import team.comit.simtong.domain.auth.dto.TokenResponse
 import team.comit.simtong.global.annotation.UseCase
 
 /**
