+++ simtong-application/src/main/kotlin/team/comit/simtong/domain/user/spi/UserJwtPort.kt
@@ -1,6 +1,6 @@
 package team.comit.simtong.domain.user.spi
 
-import team.comit.simtong.domain.auth.usecase.dto.TokenResponse
+import team.comit.simtong.domain.auth.dto.TokenResponse
 import team.comit.simtong.domain.user.model.Authority
 import java.util.UUID
