+++ simtong-application/src/main/kotlin/team/comit/simtong/domain/user/usecase/FindEmployeeNumberUseCase.kt
@@ -2,7 +2,7 @@ package team.comit.simtong.domain.user.usecase
 
 import team.comit.simtong.domain.user.exception.UserNotFoundException
 import team.comit.simtong.domain.user.spi.QueryUserPort
-import team.comit.simtong.domain.user.usecase.dto.FindEmployeeNumberRequest
+import team.comit.simtong.domain.user.dto.FindEmployeeNumberRequest
 import team.comit.simtong.global.annotation.UseCase
 
 /**
