+++ simtong-application/src/main/kotlin/team/comit/simtong/domain/user/usecase/SignInUseCase.kt
@@ -1,13 +1,13 @@
 package team.comit.simtong.domain.user.usecase
 
-import team.comit.simtong.domain.auth.usecase.dto.TokenResponse
+import team.comit.simtong.domain.auth.dto.TokenResponse
 import team.comit.simtong.domain.user.exception.DifferentPasswordException
 import team.comit.simtong.domain.user.exception.UserNotFoundException
 import team.comit.simtong.domain.user.model.Authority
 import team.comit.simtong.domain.user.spi.QueryUserPort
 import team.comit.simtong.domain.user.spi.UserJwtPort
 import team.comit.simtong.domain.user.spi.UserSecurityPort
-import team.comit.simtong.domain.user.usecase.dto.SignInRequest
+import team.comit.simtong.domain.user.dto.SignInRequest
 import team.comit.simtong.global.annotation.UseCase
 
 /**
